requirements-conda.txt
attrs==25.1.0
better-profanity==0.7.0
boto3==1.35.99
decord==0.6.0
diffusers==0.33.1
einops==0.8.1
ftfy==6.3.1
fvcore==0.1.5.post20221221
huggingface-hub==0.32.4
hydra-core==1.3.2
imageio[pyav,ffmpeg]==2.37.0
iopath==0.1.10
ipdb==0.13.13
loguru==0.7.2
mediapy==1.2.2
megatron-core==0.10.0
modelscope==1.26.0
nltk==3.9.1
numpy==1.26.4
nvidia-ml-py==12.535.133
omegaconf==2.3.0
opencv-python==4.10.0.84
pandas==2.2.3
peft==0.14.0
pillow==11.1.0
protobuf==4.25.3
pycocotools==2.0.10
pynvml==12.0.0
pyyaml==6.0.2
qwen-vl-utils[decord]==0.0.11
retinaface-py==0.0.2
safetensors==0.5.3
scikit-image==0.25.2
sentencepiece==0.2.0
termcolor==2.5.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.66.5
transformers==4.51.3
webdataset==0.2.111
